+++ tests/mouse_wheel.rs
@@ -19,20 +19,13 @@ impl ButtonlikeTestAction {
     }
 }
 
-#[derive(Clone, Copy, Debug, Reflect, PartialEq, Eq, Hash)]
+#[derive(Actionlike, Clone, Copy, Debug, Reflect, PartialEq, Eq, Hash)]
+#[actionlike(Axis)]
 enum AxislikeTestAction {
     X,
     Y,
+    #[actionlike(DualAxis)]
     XY,
-}
-
-impl Actionlike for AxislikeTestAction {
-    fn input_control_kind(&self) -> InputControlKind {
-        match self {
-            AxislikeTestAction::X | AxislikeTestAction::Y => InputControlKind::Axis,
-            AxislikeTestAction::XY => InputControlKind::DualAxis,
-        }
-    }
 }
 
 fn test_app() -> App {