+++ RELEASES.md
@@ -8,6 +8,10 @@
 
 - Reflect `Component` and `Resource`, which enables accessing the data in the type registry
 
+#### Actionlike
+
+ - added `#[actionlike]` for actions to set their input kinds, either on an enum or on its individual variants.
+
 #### ActionState
 
 - Reflect `Component` and `Resource`, which enables accessing the data in the type registry